git-svn-id: svn://192.168.1.10/public/CoffeeMud@10984 0d6f1817-ed0e-0410-87c9-987e46238f29

+++ com/planet_ink/coffee_mud/Areas/StdSpaceShip.java
@@ -129,9 +129,9 @@ public class StdSpaceShip implements Area, SpaceShip
 		return new BoundedObject.BoundedCube(coordinates(),radius());
 	}
 	
-	public int getMass()
+	public long getMass()
 	{
-		int mass=this.mass;
+		long mass=this.mass;
 		if(mass<0)
 		{
 			int newMass=phyStats().weight();
