+++ com/planet_ink/coffee_mud/Items/BasicTech/GenSpaceShip.java
@@ -202,7 +202,7 @@ public class GenSpaceShip extends StdPortal implements Electronics, SpaceShip, P
 	@Override public String getManufacturerName() { return manufacturer; }
 	@Override public void setManufacturerName(String name) { cachedManufact=null; if(name!=null) manufacturer=name; }
 	
-	@Override public int getMass()
+	@Override public long getMass()
 	{ 
 		return basePhyStats().weight()+((area instanceof SpaceShip)?((SpaceShip)area).getMass(): 1000); 
 	}
